reorganize + server tests
